Fixed a problem with different units and lattice spacings.
For a general tetragonal lattice the lattice unit factor for the Gauss constraint does not really make sense as we are combining all spatial components of the electric field, which all carry different lattice unit factors.
I corrected the Gauss constraint calculation and expressed it in physical units, which in turn has an effect on the units we use for the current and the charge density (both are now also given in physical units).
This is not really a satisfactory solution and we should think about this a little bit more.